Add: Sale

+++ PhoneShop/src/main/java/com/product/phoneshop/dto/PriceDTO.java
@@ -10,5 +10,5 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @NoArgsConstructor
 public class PriceDTO {
-    private Double salePrice;
+    private BigDecimal salePrice;
 }
