+++ PhoneShop/src/main/java/com/product/phoneshop/dto/ProductDisplayDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -13,5 +15,5 @@ public class ProductDisplayDTO {
     private String name;
     private String model;
     private String color;
-    private Double salePrice;
+    private BigDecimal salePrice;
 }
